Group parse-command type definitions
