+++ lib/types/index.ts
@@ -1,2 +1,1 @@
 export * from './dictionary'
-export * from './parsed-command'
