--- lib/types/parsed-command.ts
@@ -1,7 +1,0 @@
-import { Positional } from './positional'
-
-export interface ParsedCommand {
-  cmd: string
-  demanded: Positional[]
-  optional: Positional[]
-}
