--- lib/types/positional.ts
@@ -1,4 +1,0 @@
-export interface Positional {
-  cmd: string[]
-  variadic: boolean
-}
